feat: Tournament 타입 정의 추가 및 Match 구조 설계
