+++ src/api/types/game.ts
@@ -1,3 +1,22 @@
+export type TournamentRoundType = 'ROUND_2' | 'ROUND_4';
+
+export type Player = {
+  id: string;
+  name: string;
+  avatarUrl: string;
+  win: number;
+  lose: number;
+  tournament: number;
+};
+
+export type Match = {
+  id: string;
+  player1?: Player;
+  player2?: Player;
+  winnerId?: string;
+  children?: [Match, Match];
+};
+
 export type TournamentGame = {
   tournament_id: number;
   total_rounds: number;
@@ -7,5 +26,3 @@ export type TournamentGame = {
 export type TournamentGameList = {
   game_list: TournamentGame[];
 };
-
-export type TournamentRoundType = 'ROUND_2' | 'ROUND_4';
