+++ src/pages/GamePage/TournamentPage/components/Chat/styles.css.ts
