+++ src/pages/GamePage/TournamentPage/components/MatchNode/index.tsx
