+++ src/pages/GamePage/TournamentPage/components/MatchNode/styles.css.ts
