+++ src/pages/GamePage/TournamentPage/components/PlayerCard/index.tsx
