+++ src/pages/GamePage/TournamentPage/styles.css.ts
